progress on refactoring and making tests pass
